[Native-IO] Default writes to vortex compact
Set the default physical format to vortex-compact across Rust, Python, Spark, and Flink native write paths.

Rename Spark and Flink native writer classes to LakeSoul-neutral names while keeping deprecated Parquet-named compatibility wrappers.
