+++ lakesoul-spark/src/main/scala/org/apache/spark/sql/lakesoul/Snapshot.scala
@@ -8,7 +8,7 @@ import com.dmetasoul.lakesoul.meta.{DataFileInfo, PartitionInfoScala, SparkMetaV
 import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.execution.datasources.FileFormat
 import org.apache.spark.sql.execution.datasources.parquet.ParquetFileFormat
-import org.apache.spark.sql.execution.datasources.v2.parquet.NativeParquetFileFormat
+import org.apache.spark.sql.execution.datasources.v2.lakesoul.NativeLakeSoulFileFormat
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.lakesoul.LakeSoulOptions.ReadType
 import org.apache.spark.sql.lakesoul.sources.LakeSoulSQLConf
@@ -71,7 +71,7 @@ class Snapshot(table_info: TableInfo,
 
   /** Return the underlying Spark `FileFormat` of the LakeSoulTableRel. */
   def fileFormat: FileFormat = if (SQLConf.get.getConf(LakeSoulSQLConf.NATIVE_IO_ENABLE)) {
-    new NativeParquetFileFormat()
+    new NativeLakeSoulFileFormat()
   } else {
     new ParquetFileFormat()
   }
